adding screen transition using a global game_state
